Checkpoint: Exercise 31 in progress
There is an open exercise to add some improvement to
movie-similarities.py

Choosing to write an algorithm that first filters out movie pairs that
are not related by at least one genre, the employees.py script attempts
a simplified version of the main challenge from my approach. Generating
a "reference RDD", and checking the original RDD against that reference
while filtering out any element of the original RDD whose key does not
exist in the reference RDD.

Intersection.py attempts to "scale up" the problem to the level of the
movie-similarities script, and is technically successful. However,
when performing any sort of "reduce" function, such as .count(), then
the issue becomes clear that the algorithm is not optimized.

The movie-similarities.py can be left to run for over 5 hours without
any result. This will stay like this for now. Perhaps with wisdom
achieved later on, I can learn how to efficiently use some sort of
"reference RDD"